+++ test/unit/range/views/as_const_test.cpp
@@ -16,7 +16,6 @@
 #include <seqan3/range/concept.hpp>
 #include <seqan3/range/views/as_const.hpp>
 #include <seqan3/range/views/complement.hpp>
-#include <seqan3/range/views/to_lower.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
 
 using seqan3::operator""_dna5;
@@ -40,7 +39,7 @@ TEST(view_as_const, basic)
 
 TEST(view_as_const, concepts)
 {
-    std::string vec{"ACTTTGATA"};
+    seqan3::dna5_vector vec{"ACTTTGATA"_dna5};
     auto v1 = vec | seqan3::views::as_const;
     EXPECT_TRUE(std::ranges::input_range<decltype(v1)>);
     EXPECT_TRUE(std::ranges::forward_range<decltype(v1)>);
@@ -52,18 +51,19 @@ TEST(view_as_const, concepts)
     EXPECT_TRUE(seqan3::const_iterable_range<decltype(v1)>);
     EXPECT_FALSE((std::ranges::output_range<decltype(v1), char>));
 
-    EXPECT_TRUE((std::is_same_v<decltype(v1[0]), char const &>));
+    EXPECT_TRUE((std::is_same_v<decltype(v1[0]), seqan3::dna5 const &>));
 
-    auto v2 = vec | seqan3::views::to_lower | seqan3::views::as_const; // to_lower generates values
-    EXPECT_TRUE(std::ranges::input_range<decltype(v1)>);
-    EXPECT_TRUE(std::ranges::forward_range<decltype(v1)>);
-    EXPECT_TRUE(std::ranges::bidirectional_range<decltype(v1)>);
-    EXPECT_TRUE(std::ranges::random_access_range<decltype(v1)>);
-    EXPECT_TRUE(std::ranges::view<decltype(v1)>);
-    EXPECT_TRUE(std::ranges::sized_range<decltype(v1)>);
-    EXPECT_TRUE(std::ranges::common_range<decltype(v1)>);
-    EXPECT_TRUE(seqan3::const_iterable_range<decltype(v1)>);
-    EXPECT_FALSE((std::ranges::output_range<decltype(v1), char>));
+    // complement generates values
+    auto v2 = vec | seqan3::views::complement | seqan3::views::as_const;
+    EXPECT_TRUE(std::ranges::input_range<decltype(v2)>);
+    EXPECT_TRUE(std::ranges::forward_range<decltype(v2)>);
+    EXPECT_TRUE(std::ranges::bidirectional_range<decltype(v2)>);
+    EXPECT_TRUE(std::ranges::random_access_range<decltype(v2)>);
+    EXPECT_TRUE(std::ranges::view<decltype(v2)>);
+    EXPECT_TRUE(std::ranges::sized_range<decltype(v2)>);
+    EXPECT_TRUE(std::ranges::common_range<decltype(v2)>);
+    EXPECT_TRUE(seqan3::const_iterable_range<decltype(v2)>);
+    EXPECT_FALSE((std::ranges::output_range<decltype(v2), char>));
 
-    EXPECT_TRUE((std::is_same_v<decltype(v2[0]), char>)); // don't add const-ness to values
+    EXPECT_TRUE((std::is_same_v<decltype(v2[0]), seqan3::dna5>)); // don't add const-ness to values
 }