+++ test/unit/range/views/to_lower_test.cpp
@@ -15,6 +15,10 @@
 #include <seqan3/range/views/to_lower.hpp>
 #include <seqan3/test/expect_range_eq.hpp>
 
+#ifdef SEQAN3_DEPRECATED_310
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wdeprecated-declarations"
+
 using seqan3::operator""_dna5;
 
 TEST(view_to_lower, basic)
@@ -94,3 +98,5 @@ TEST(view_to_lower, concepts)
     EXPECT_FALSE((std::ranges::output_range<decltype(lower_view), char>));
     EXPECT_FALSE(std::ranges::contiguous_range<decltype(lower_view)>);
 }
+#pragma GCC diagnostic pop
+#endif // SEQAN3_DEPRECATED_310