+++ CHANGELOG.md
@@ -206,6 +206,9 @@ regression test suite and patches at https://github.com/seqan/seqan3/tree/master
 * We deprecated `seqan3::views::to_upper` and it will be removed in 3.1.0, use
   `std::views::transform([](auto && chr){return std::toupper(chr)})`.
   ([\#2540](https://github.com/seqan/seqan3/pull/2538))
+* We deprecated `seqan3::views::to_lower` and it will be removed in 3.1.0, use
+   `std::views::transform([](auto && chr){return std::tolower(chr)})`.
+   ([\#2556](https://github.com/seqan/seqan3/pull/2556))
 
 #### Search